+++ resdk/resolwe.py
@@ -75,7 +75,10 @@ class Resolwe(ResolweUtilsMixin):
         """Initialize attributes."""
         if url is None:
             # Try to get URL from environmental variable, otherwise fallback to default.
-            url = os.environ.get('RESOLWE_API_HOST', DEFAULT_URL)
+            url = os.environ.get('RESOLWE_HOST_URL', DEFAULT_URL)
+            # TODO: Remove this
+            if 'RESOLWE_HOST_URL' not in os.environ:
+                url = os.environ.get('RESOLWE_API_HOST', url)
 
         if username is None:
             username = os.environ.get('RESOLWE_API_USERNAME', None)